Add icons and css
