--- client/my-sites/email/inbox/mailbox-selection-list/style.scss
@@ -1,13 +1,0 @@
-@import '@wordpress/base-styles/breakpoints';
-@import '@wordpress/base-styles/mixins';
-
-.mailbox-selection-list__header {
-	.formatted-header__title {
-		font-size: $font-headline-medium;
-		line-height: $font-headline-medium;
-	}
-	.formatted-header__subtitle {
-		font-size: $font-body;
-		line-height: $font-title-medium;
-	}
-}
